<REFACTOR> intergrate api endpoint live start and end to start & end application
